+++ packages/dds/tree/src/feature-libraries/default-schema/index.ts
@@ -26,7 +26,12 @@ export {
 	relevantRemovedRoots,
 } from "./defaultEditBuilder.js";
 
-export { SchemaValidationErrors, isNodeInSchema, isFieldInSchema } from "./schemaChecker.js";
+export {
+	SchemaValidationErrors,
+	isNodeInSchema,
+	isFieldInSchema,
+	inSchemaOrThrow,
+} from "./schemaChecker.js";
 
 export { defaultSchemaPolicy } from "./defaultSchema.js";
