+++ packages/dds/tree/src/feature-libraries/index.ts
@@ -145,6 +145,7 @@ export {
 	SchemaValidationErrors,
 	isNodeInSchema,
 	isFieldInSchema,
+	inSchemaOrThrow,
 } from "./default-schema/index.js";
 
 export {
