+++ packages/dds/tree/src/test/simple-tree/api/create.spec.ts
@@ -5,8 +5,16 @@
 
 import { strict as assert } from "node:assert";
 
-import { createFromInsertable, SchemaFactory } from "../../../simple-tree/index.js";
+import {
+	createFromInsertable,
+	cursorFromInsertable,
+	// eslint-disable-next-line import/no-internal-modules
+} from "../../../simple-tree/api/create.js";
+
 import { TreeAlpha } from "../../../shared-tree/index.js";
+import { SchemaFactory } from "../../../simple-tree/index.js";
+import { MockNodeIdentifierManager } from "../../../feature-libraries/index.js";
+import { validateUsageError } from "../../utils.js";
 
 const schema = new SchemaFactory("com.example");
 
@@ -30,5 +38,24 @@ describe("simple-tree create", () => {
 		});
 		const canvas2 = new Canvas({ stuff: [] });
 		assert.deepEqual(canvas1, canvas2);
+	});
+
+	describe("cursorFromInsertable", () => {
+		it("Success", () => {
+			cursorFromInsertable(schema.string, "Hello world", new MockNodeIdentifierManager());
+		});
+
+		it("Failure", () => {
+			assert.throws(
+				() =>
+					cursorFromInsertable(
+						schema.number,
+						// @ts-expect-error invalid data for schema
+						"Hello world",
+						new MockNodeIdentifierManager(),
+					),
+				validateUsageError(/incompatible/),
+			);
+		});
 	});
 });
